LumberjackToolkit

Documentation for LumberjackToolkit.

LumberjackToolkit._correct_center

LumberjackToolkit._correct_symmetry

LumberjackToolkit._get_reference

LumberjackToolkit._ordering_symmetry

LumberjackToolkit.average_bins

LumberjackToolkit.averaging_profile

LumberjackToolkit.binning

LumberjackToolkit.densityprofile

LumberjackToolkit.dihedral

LumberjackToolkit.select_atoms

LumberjackToolkit.simulation_steps

LumberjackToolkit.vmd_get_charges

LumberjackToolkit.δ

LumberjackToolkit.ρ

LumberjackToolkit._correct_center — Method

_correct_center(bins::Vector{Float64}, reference::Float64)

It corrects the bin values based on the referenced center. In this way, the center will become zero and all the bins values will be reajusted to apply it. Now, every distance is compared with the central value.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

reference::Float64: a Float64 with the reference center.

LumberjackToolkit._correct_symmetry — Method

_correct_symmetry(bins::Vector{Float64}, reference::Float64)

Do the same thing that _correct_center(...), but it extends the concept to adjust the data in such way that the center will become the center of the box in every frame.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

reference::Float64: a Float64 with the reference center.

LumberjackToolkit._get_reference — Method

_get_reference(simulation::MolSimToolkit.Simulation, selection::String)

It aims to get the reference center for each frame along the trajectory.

Arguments

simulation::MolSimToolkit.Simulation: a MolSimToolkit.Simulation object.

selection::String: a selection string to get the reference center. For exemple, "all" or "not water".

LumberjackToolkit._ordering_symmetry — Method

_ordering_symmetry(bins::Vector{Float64}, ρ::Vector{Float64})

This function checks if the bins are ordered after the symmetry protocol and if not, it orders them.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

ρ::Vector{Float64}: a Vector{Float64} with the densities.

LumberjackToolkit.average_bins — Method

average_bins(bins::Vector{Float64})

It takes an average vector based on the binned box parameters. So it can be rightly compare with the ρ density with $length(density) ≡ length(bins) - 1$.

LumberjackToolkit.averaging_profile — Method

averaging_profile(distances::Vector{Vector{Float64}}, densities::Vector{Vector{Float64}})

This function takes the distances and densities from Vector{Vector{Float64}} to Vector{Float64} by means and standard deviation computations.

Arguments

distances::Vector{Vector{Float64}}: a Vector{Vector{Float64}} with the distances from the center of the profile.

densities::Vector{Vector{Float64}}: a Vector{Vector{Float64}} with the densities of the profile.

LumberjackToolkit.binning — Method

binning(simulation::MolSimToolkit.Simulation; axis="z", resolution=1.)

It reports the bins and normalization factor for the loaded simulation box.

Arguments

simulation::MolSimToolkit.Simulation: a MolSimToolkit.Simulation object.

axis::String: a String with the axis flag.

dimensions::Int: an integer with the number of dimensions. It can be 1 or 2.

resolution::Float64: a Float64 with the resolution of the bins.

LumberjackToolkit.densityprofile — Method

densityprofile(pdb::String, trajectory::String; ...)

It computes the density profile for each one of the frames inside the trajectory.

Arguments

pdb::String: a String with the pdb file path.

trajectory::String: a String with the trajectory file path.

profile::String: a String with the profile type. It can be "number", "mass", "charge", or "electron".

selection::String: a String with the selection of the atoms.

axis::String: a String with the axis flag.

dimensions::Int: an integer with the number of dimensions. It can be 1 or 2.

resolution::Float64: a Float64 with the resolution of the bins.

center::Bool: a Bool to center the profile.

symmetry::Bool: a Bool to apply the symmetry protocol.

symmetry_selection::String: a String with the selection of the atoms to apply the symmetry protocol.

first::Int: an integer with the first frame to be considered.

step::Int: an integer with the step to be considered.

last::Int: an integer with the last frame to be considered.

charge_column::String: a String with the charge column to be considered.

edp::String: a String with the electron density profile.

charge_correction::Bool: a Bool to apply the charge correction.

normalization::Bool: a Bool to apply the normalization.

echo::Bool: a Bool to print or not the output.

LumberjackToolkit.dihedral — Method

dihedral(atom1::StaticArrays.SVector, atom2::StaticArrays.SVector, atom3::StaticArrays.SVector, atom4::StaticArrays.SVector)

Calculate the dihedral angle between four atoms. The dihedral angle is the angle between the planes defined by the atoms (atom1, atom2, atom3) and (atom2, atom3, atom4). The dihedral angle is calculated using the atan method.

LumberjackToolkit.select_atoms — Method

select_atoms(atom::Vector{PDBTools.Atom}, name::String, resname::String, resnum::Int64, segname::String)

Select the atoms that match the given name, residue name, residue number, and segment name.

LumberjackToolkit.simulation_steps — Method

simulation_steps(t_simulation::Int64; unit="ns", timestep=2.0, output=1000)

This function calculates the number of steps you need to set and the number of frames that you will get in the MD simulation.

Arguments

t_simulation: The total simulation time in the specified unit.

unit: The unit of the simulation time. It can be "ns" (nan

timestep: The time step of the simulation in the specified unit.

output: The frequency of frame storage during the MD simulation.

LumberjackToolkit.vmd_get_charges — Method

vmd_get_charges(psf::String, pdb::String; newpdb=nothing, pdb_column="beta", vmd="vmd")

Uses VMD to catch the charges inside the PSF file and transfer it to PDB file on b-factor or occupancy columns. Notice that you need VMD installed to run this code. Check it on: https://www.ks.uiuc.edu/Development/Download/download.cgi?PackageName=VMD.

Arguments

psf::String and pdb::String: the name of the raw files.

newpdb::String: the name of the new PDB file with the charges. The default is nothing and it will generate a PDB on your temporary files.

pdb_column::String: the column to store the charges. The default is beta.

vmd::String: the VMD executable. The default is vmd.

LumberjackToolkit.δ — Method

δ(value::Float64, lower::Float64, upper::Float64)

It is an indicator function that maps the elements inside $x_lower ≤ x_value ≤ x_upper$ interval.

LumberjackToolkit.ρ — Method

ρ(bins::Vector{Float64}, positions::Vector{Point3D{Float64}}, N::Float64; axis="z", prop=nothing)

Computes the density for each bin interval given the frame positions. Here, we apply a normalization base on the bin dimensions, so all the frames will have the same sum(ρ). The prop argument is used to compute the density of a specific property, like the atomic number, mass, or charge.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bin intervals.

positions::Vector{Point3D{Float64}}: a Vector{Point3D{Float64}} with the atomic positions.

N::Float64: a Float64 with the normalization factor.

axis::String: a String with the axis flag.

prop::Vector{Float64}: a Vector{Float64} with the property to be computed.