LumberjackToolkit

Documentation for LumberjackToolkit.

LumberjackToolkit._correct_center

LumberjackToolkit._correct_symmetry

LumberjackToolkit._get_reference

LumberjackToolkit._ordering_symmetry

LumberjackToolkit.align_trajectory

LumberjackToolkit.average_bins

LumberjackToolkit.averaging_profile

LumberjackToolkit.binning

LumberjackToolkit.cellulose_surface

LumberjackToolkit.densityprofile

LumberjackToolkit.dihedral

LumberjackToolkit.dihedrals

LumberjackToolkit.dispm

LumberjackToolkit.lovo_mapping

LumberjackToolkit.msd

LumberjackToolkit.r_avg

LumberjackToolkit.readcoords

LumberjackToolkit.select_atoms

LumberjackToolkit.simsteps

LumberjackToolkit.simulation_steps

LumberjackToolkit.vmd_get_charges

LumberjackToolkit.write_frame!

LumberjackToolkit.writepdb_trajectory

LumberjackToolkit.δ

LumberjackToolkit.ρ

LumberjackToolkit._correct_center — Method

_correct_center(bins::Vector{Float64}, reference::Float64)

It corrects the bin values based on the referenced center. In this way, the center will become zero and all the bins values will be reajusted to apply it. Now, every distance is compared with the central value.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

reference::Float64: a Float64 with the reference center.

LumberjackToolkit._correct_symmetry — Method

_correct_symmetry(bins::Vector{Float64}, reference::Float64)

Do the same thing that _correct_center(...), but it extends the concept to adjust the data in such way that the center will become the center of the box in every frame.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

reference::Float64: a Float64 with the reference center.

LumberjackToolkit._get_reference — Method

_get_reference(simulation::MolSimToolkit.Simulation, selection::String)

It aims to get the reference center for each frame along the trajectory.

Arguments

simulation::MolSimToolkit.Simulation: a MolSimToolkit.Simulation object.

selection::String: a selection string to get the reference center. For exemple, "all" or "not water".

LumberjackToolkit._ordering_symmetry — Method

_ordering_symmetry(bins::Vector{Float64}, ρ::Vector{Float64})

This function checks if the bins are ordered after the symmetry protocol and if not, it orders them.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bins.

ρ::Vector{Float64}: a Vector{Float64} with the densities.

LumberjackToolkit.align_trajectory — Method

align_trajectory(psfname::String, pdbname::String, trajectory::String; ...)

Aligns a trajectory to a reference structure using VMD. It loads the psf, pdb, and trajectory files, wraps the trajectory, and aligns the frames to the reference structure. The selection argument is used to define the atoms to be aligned based on the VMD selection, such as a protein or a cellulose fibril. The DebugVMD argument equals true will return the VMD output. The function returns the path to the new trajectory. If the new_trajectory argument is not provided, a temporary file is created.

Arguments

psfname::String: The path to the PSF file.

pdbname::String: The path to the PDB file.

trajectory::String: The path to the trajectory file.

selection::String: The VMD selection to be aligned. Default is "protein".

pbc_dimensions::String: The PBC dimensions to be set on pbctools format. If not provided, the function will try to read the dimensions from the .xsc file, based on the location of trajectory.

new_trajectory::String: The path to the new trajectory file. If not provided, a temporary file is created.

reference::Int: The frame to be used as reference. Default is 0, the initial frame.

vmd::String: The path to the VMD executable. Default is "vmd".

DebugVMD::Bool: If true, the function will return the VMD output. Default is false.

LumberjackToolkit.average_bins — Method

average_bins(bins::Vector{Float64})

It takes an average vector based on the binned box parameters. So it can be rightly compare with the ρ density with $length(density) ≡ length(bins) - 1$.

LumberjackToolkit.averaging_profile — Method

averaging_profile(distances::Vector{Vector{Float64}}, densities::Vector{Vector{Float64}})

This function takes the distances and densities from Vector{Vector{Float64}} to Vector{Float64} by means and standard deviation computations.

Arguments

distances::Vector{Vector{Float64}}: a Vector{Vector{Float64}} with the distances from the center of the profile.

densities::Vector{Vector{Float64}}: a Vector{Vector{Float64}} with the densities of the profile.

LumberjackToolkit.binning — Method

binning(simulation::MolSimToolkit.Simulation; axis="z", resolution=1.)

It reports the bins and normalization factor for the loaded simulation box.

Arguments

simulation::MolSimToolkit.Simulation: a MolSimToolkit.Simulation object.

axis::String: a String with the axis flag.

dimensions::Int: an integer with the number of dimensions. It can be 1 or 2.

resolution::Float64: a Float64 with the resolution of the bins.

LumberjackToolkit.cellulose_surface — Method

surface(rawfile::String; surface_method="distance", selection="all", parameters="-res 0.6 -cutoff 4.0", vmd="vmd")

Generate a surface from a VMD input file using the volmap plugin. The surface can be generated using different methods, such as distance, density, among others. The default method is distance. The selection can be defined by the user, and the default is all.

The parameters can be set by the user, and the default is -res 0.6 -cutoff 4.0. The VMD executable can be provided as an argument whether is not on default PATH.

Arguments

rawfile::String: The name of the input file without extension. The script will look for rawfile.pdb and rawfile.psf.

surface_method::String="distance": The method to be used to generate the surface using the volmap tools. The default is distance. density creates a map of the weighted atomic density at each gridpoint. This is done by replacing each atom in the selection with a normalized gaussian distribution of width (standard deviation) equal to its atomic radius. The gaussian distribution for each atom is then weighted using an optional -weight argument, and defaults to a weight of one (i.e, the number density). The various gaussians are then additively distributed on a grid. interp creates a map with the atomic weights interpolated onto a grid. For each atom, its weight is distributed to the 8 nearest voxels via a trilinear interpolation. The optional -weight argument defaults to a weight of one. distance creates a map for which each gridpoint contains the distance between that point and the edge of the nearest atom. In other words, each gridpoint specifies the maximum radius of a sphere cnetered at that point which does not intersect with the spheres of any other atoms. All atoms are treated as spheres using the atoms' VMD radii. coulomb, coulombmsm Creates a map of the electrostatic field of the atom selection, made by computing the non-bonded Coulomb potential from each atom in the selection (in units of k_B.T/e). The coulomb map generation is optimized to take advantage of multi-core CPUs and programmable GPUs if they are available. ils Creates a free energy map of the distribution of a weakly-interacting monoatomic or diatomic gas ligand throughout the system using the Implicit Ligand Sampling (ILS) technique. See additional information about ILS below. mask Creates a map which is set to 0 or 1 depending on whether they are within a specified -cutoff distance argument of any atoms in the selection. The mask map is typically used in combination with other maps in order to hide/mask data that is far from a region of interest. occupancy Each grid point is set to either 0 or 1, depending on whether it contains onbe or more atoms or not. When averaged over many frames, this will provide the fractional occupancy of that grid point. By default, atoms are treated as spheres using the atomic radii and a gridpoint is considered to be "occupied" if it lies inside that sphere. Use the -points argument to treat atoms as points (a grid point is "occupied" if its grid cube contains an atom's center).

selection::String="all": The selection to be used to generate the surface. The default is all.

parameters::String="-res 0.6 -cutoff 4.0": The parameters to be used in the volmap plugin. The default is -res 0.6 -cutoff 4.0. To understand the parameters, see the VMD documentation https://www.ks.uiuc.edu/Research/vmd/vmd-1.9.4/ug/node157.html.

vmd::String="vmd": The VMD executable. The default is vmd.

Examples

julia > surface("new_crystal")

LumberjackToolkit.densityprofile — Method

densityprofile(pdb::String, trajectory::String; ...)

It computes the density profile for each one of the frames inside the trajectory.

Arguments

pdb::String: a String with the pdb file path.

trajectory::String: a String with the trajectory file path.

profile::String: a String with the profile type. It can be "number", "mass", "charge", or "electron".

selection::String: a String with the selection of the atoms.

axis::String: a String with the axis flag.

dimensions::Int: an integer with the number of dimensions. It can be 1 or 2.

resolution::Float64: a Float64 with the resolution of the bins.

center::Bool: a Bool to center the profile.

symmetry::Bool: a Bool to apply the symmetry protocol.

symmetry_selection::String: a String with the selection of the atoms to apply the symmetry protocol.

first::Int: an integer with the first frame to be considered.

step::Int: an integer with the step to be considered.

last::Int: an integer with the last frame to be considered.

charge_column::String: a String with the charge column to be considered.

edp::String: a String with the electron density profile.

charge_correction::Bool: a Bool to apply the charge correction.

normalization::Bool: a Bool to apply the normalization.

echo::Bool: a Bool to print or not the output.

LumberjackToolkit.dihedral — Method

dihedral(atom1::StaticArrays.SVector, atom2::StaticArrays.SVector, atom3::StaticArrays.SVector, atom4::StaticArrays.SVector)

Calculate the dihedral angle between four atoms. The dihedral angle is the angle between the planes defined by the atoms (atom1, atom2, atom3) and (atom2, atom3, atom4). The dihedral angle is calculated using the atan method.

LumberjackToolkit.dihedrals — Method

dihedrals(pdbname::String, trajectory::String, segname::String; ...)

Calculate the dihedral angles and their sum based on the carbohydrate chain. The pdbname carries the atom information, and the trajectory holds the coordinates for each atoms in the simulation. The carbohydrate chain is defined by the segname parameter, that is the same segname present on PDB.

Wohlert uses a lot the ϕ(O5'-C1'-O4-C4) and ψ(C1'-O4-C4-C3) definition, while Carol uses ϕ(O5'-C1'-O4-C4) and ψ(C1'-O4-C4-C5)

LumberjackToolkit.dispm — Method

dispm(frames::Vector{<:Vector{SVector{3, Float64}}}, weights::Vector{Float64}; weight_method=nothing)

Calculate the displacement matrix of a trajectory with weights.

LumberjackToolkit.lovo_mapping — Method

map_fractions(atoms::AbstractVector{<:PDBTools.Atom}, trajectory_file::String)

LumberjackToolkit.msd — Method

msd(trajectory::Vector{Vector{SVector{3, Float64}}}; step=1, dim=collect(1:3))

Calculate the mean square displacement (MSD) of a trajectory. The MSD is calculated using Green-Kubo relation: <r²(t)> = MSD(t) = 1/N * Σ <[r_i(t) - r_i(0)]²>. It is important to notice that trajectories must have been adjusted to approach:

    traj = [
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 1
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 2
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 3
                ...
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule N
        ]

Arguments

trajectory::Vector{Vector{SVector{3, Float64}}}: A vector of vectors of 3D coordinates.

step::Int=1: The time step to calculate the MSD.

dim::Vector{Int}=collect(1:3): The dimensions to calculate the MSD.

LumberjackToolkit.r_avg — Method

r_avg(frames::Vector{<:Vector{SVector{3, Float64}}})

Calculate the average position of each atom inside the trajectory.

LumberjackToolkit.readcoords — Method

readcoords(atoms::AbstractVector{<:PDBTools.Atom}, pdbname::String)

Read the coordinates of a trajectory in a PDB file. Returns a vector of vectors of coordinates of selected atoms.

LumberjackToolkit.select_atoms — Method

select_atoms(atom::Vector{PDBTools.Atom}, name::String, resname::String, resnum::Int64, segname::String)

Select the atoms that match the given name, residue name, residue number, and segment name.

LumberjackToolkit.simsteps — Method

LumberjackToolkit.simulation_steps — Method

simulation_steps(t_simulation::Int64; unit="ns", timestep=2.0, output=1000)

This function calculates the number of steps you need to set and the number of frames that you will get in the MD simulation.

Arguments

t_simulation: The total simulation time in the specified unit.

unit: The unit of the simulation time. It can be "ns" (nan

timestep: The time step of the simulation in the specified unit.

output: The frequency of frame storage during the MD simulation.

LumberjackToolkit.vmd_get_charges — Method

vmd_get_charges(psf::String, pdb::String; newpdb=nothing, pdb_column="beta", vmd="vmd")

Uses VMD to catch the charges inside the PSF file and transfer it to PDB file on b-factor or occupancy columns. Notice that you need VMD installed to run this code. Check it on: https://www.ks.uiuc.edu/Development/Download/download.cgi?PackageName=VMD.

Arguments

psf::String and pdb::String: the name of the raw files.

newpdb::String: the name of the new PDB file with the charges. The default is nothing and it will generate a PDB on your temporary files.

pdb_column::String: the column to store the charges. The default is beta.

vmd::String: the VMD executable. The default is vmd.

LumberjackToolkit.write_frame! — Method

write_frame!(trajectory_pdb_file, atoms, frame

Write a frame of a trajectory in the temporary PDB trajectory file.

LumberjackToolkit.writepdb_trajectory — Method

writepdb_trajectory(pdbname::String, trajectory::String; selection="all", first=1, last=nothing, step=1)

Write a trajectory in a temporary PDB file. This is very useful when you want to visualize the edited trajectory in VMD. The MDLovoFit program uses this function to write the trajectory in a PDB file before the analysis.

LumberjackToolkit.δ — Method

δ(value::Float64, lower::Float64, upper::Float64)

It is an indicator function that maps the elements inside $x_lower ≤ x_value ≤ x_upper$ interval.

LumberjackToolkit.ρ — Method

ρ(bins::Vector{Float64}, positions::Vector{Point3D{Float64}}, N::Float64; axis="z", prop=nothing)

Computes the density for each bin interval given the frame positions. Here, we apply a normalization base on the bin dimensions, so all the frames will have the same sum(ρ). The prop argument is used to compute the density of a specific property, like the atomic number, mass, or charge.

Arguments

bins::Vector{Float64}: a Vector{Float64} with the bin intervals.

positions::Vector{Point3D{Float64}}: a Vector{Point3D{Float64}} with the atomic positions.

N::Float64: a Float64 with the normalization factor.

axis::String: a String with the axis flag.

prop::Vector{Float64}: a Vector{Float64} with the property to be computed.