LumberjackToolkit

Documentation for LumberjackToolkit.

LumberjackToolkit.align_frames

LumberjackToolkit.cellulose_surface

LumberjackToolkit.chargesPSF

LumberjackToolkit.dihedral

LumberjackToolkit.dihedrals

LumberjackToolkit.dispm

LumberjackToolkit.lovo_mapping

LumberjackToolkit.msd

LumberjackToolkit.r_avg

LumberjackToolkit.readcoords

LumberjackToolkit.select_atoms

LumberjackToolkit.simsteps

LumberjackToolkit.simulation_steps

LumberjackToolkit.write_frame!

LumberjackToolkit.writepdb_trajectory

LumberjackToolkit.align_frames — Method

vmd_trajectory_alignment(psfname::String, trjname::String; new_trajectory=nothing, selection="not water", reference=0, vmd="vmd", DebugVMD=false)

Align the frames of a trajectory using the VMD measure fit command. The selection can be defined by the user, and the default is not water. The reference frame can be defined by the user, and the default is 0.

Arguments

psfname::String: The name of the PSF file.

trjname::String: The name of the trajectory file. It can be in any format that VMD can read.

new_trajectory::String=nothing: The name of the new DCD trajectory file. The default creates a temporary file.

selection::String="not water": The selection to be used to align the frames.

reference=0: The reference frame to be used to align the frames.

vmd="vmd": The VMD executable. The default is vmd.

DebugVMD=false: If true, the output of VMD will be printed.

LumberjackToolkit.cellulose_surface — Method

surface(rawfile::String; surface_method="distance", selection="all", parameters="-res 0.6 -cutoff 4.0", vmd="vmd")

Generate a surface from a VMD input file using the volmap plugin. The surface can be generated using different methods, such as distance, density, among others. The default method is distance. The selection can be defined by the user, and the default is all.

The parameters can be set by the user, and the default is -res 0.6 -cutoff 4.0. The VMD executable can be provided as an argument whether is not on default PATH.

Arguments

rawfile::String: The name of the input file without extension. The script will look for rawfile.pdb and rawfile.psf.

surface_method::String="distance": The method to be used to generate the surface using the volmap tools. The default is distance. density creates a map of the weighted atomic density at each gridpoint. This is done by replacing each atom in the selection with a normalized gaussian distribution of width (standard deviation) equal to its atomic radius. The gaussian distribution for each atom is then weighted using an optional -weight argument, and defaults to a weight of one (i.e, the number density). The various gaussians are then additively distributed on a grid. interp creates a map with the atomic weights interpolated onto a grid. For each atom, its weight is distributed to the 8 nearest voxels via a trilinear interpolation. The optional -weight argument defaults to a weight of one. distance creates a map for which each gridpoint contains the distance between that point and the edge of the nearest atom. In other words, each gridpoint specifies the maximum radius of a sphere cnetered at that point which does not intersect with the spheres of any other atoms. All atoms are treated as spheres using the atoms' VMD radii. coulomb, coulombmsm Creates a map of the electrostatic field of the atom selection, made by computing the non-bonded Coulomb potential from each atom in the selection (in units of k_B.T/e). The coulomb map generation is optimized to take advantage of multi-core CPUs and programmable GPUs if they are available. ils Creates a free energy map of the distribution of a weakly-interacting monoatomic or diatomic gas ligand throughout the system using the Implicit Ligand Sampling (ILS) technique. See additional information about ILS below. mask Creates a map which is set to 0 or 1 depending on whether they are within a specified -cutoff distance argument of any atoms in the selection. The mask map is typically used in combination with other maps in order to hide/mask data that is far from a region of interest. occupancy Each grid point is set to either 0 or 1, depending on whether it contains onbe or more atoms or not. When averaged over many frames, this will provide the fractional occupancy of that grid point. By default, atoms are treated as spheres using the atomic radii and a gridpoint is considered to be "occupied" if it lies inside that sphere. Use the -points argument to treat atoms as points (a grid point is "occupied" if its grid cube contains an atom's center).

selection::String="all": The selection to be used to generate the surface. The default is all.

parameters::String="-res 0.6 -cutoff 4.0": The parameters to be used in the volmap plugin. The default is -res 0.6 -cutoff 4.0. To understand the parameters, see the VMD documentation https://www.ks.uiuc.edu/Research/vmd/vmd-1.9.4/ug/node157.html.

vmd::String="vmd": The VMD executable. The default is vmd.

Examples

julia > surface("new_crystal")

LumberjackToolkit.chargesPSF — Method

chargesPSF(psfname::String)

Extract the partial charges from a PSF file.

LumberjackToolkit.dihedral — Method

dihedral(atom1::StaticArrays.SVector, atom2::StaticArrays.SVector, atom3::StaticArrays.SVector, atom4::StaticArrays.SVector)

Calculate the dihedral angle between four atoms. The dihedral angle is the angle between the planes defined by the atoms (atom1, atom2, atom3) and (atom2, atom3, atom4). The dihedral angle is calculated using the atan method.

LumberjackToolkit.dihedrals — Method

dihedrals(pdbname::String, trajectory::String, segname::String; ...)

Calculate the dihedral angles and their sum based on the carbohydrate chain. The pdbname carries the atom information, and the trajectory holds the coordinates for each atoms in the simulation. The carbohydrate chain is defined by the segname parameter, that is the same segname present on PDB.

Wohlert uses a lot the ϕ(O5'-C1'-O4-C4) and ψ(C1'-O4-C4-C3) definition, while Carol uses ϕ(O5'-C1'-O4-C4) and ψ(C1'-O4-C4-C5)

LumberjackToolkit.dispm — Method

dispm(frames::Vector{<:Vector{SVector{3, Float64}}}, weights::Vector{Float64}; weight_method=nothing)

Calculate the displacement matrix of a trajectory with weights.

LumberjackToolkit.lovo_mapping — Method

map_fractions(atoms::AbstractVector{<:PDBTools.Atom}, trajectory_file::String)

LumberjackToolkit.msd — Method

msd(trajectory::Vector{Vector{SVector{3, Float64}}}; step=1, dim=collect(1:3))

Calculate the mean square displacement (MSD) of a trajectory. The MSD is calculated using Green-Kubo relation: <r²(t)> = MSD(t) = 1/N * Σ <[r_i(t) - r_i(0)]²>. It is important to notice that trajectories must have been adjusted to approach:

    traj = [
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 1
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 2
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule 3
                ...
                [coord(t1), coord(t2), coord(t3), ..., coord(tmax)],    ## molecule N
        ]

Arguments

trajectory::Vector{Vector{SVector{3, Float64}}}: A vector of vectors of 3D coordinates.

step::Int=1: The time step to calculate the MSD.

dim::Vector{Int}=collect(1:3): The dimensions to calculate the MSD.

LumberjackToolkit.r_avg — Method

r_avg(frames::Vector{<:Vector{SVector{3, Float64}}})

Calculate the average position of each atom inside the trajectory.

LumberjackToolkit.readcoords — Method

readcoords(atoms::AbstractVector{<:PDBTools.Atom}, pdbname::String)

Read the coordinates of a trajectory in a PDB file. Returns a vector of vectors of coordinates of selected atoms.

LumberjackToolkit.select_atoms — Method

select_atoms(atom::Vector{PDBTools.Atom}, name::String, resname::String, resnum::Int64, segname::String)

Select the atoms that match the given name, residue name, residue number, and segment name.

LumberjackToolkit.simsteps — Method

LumberjackToolkit.simulation_steps — Method

simulation_steps(t_simulation::Int64; unit="ns", timestep=2.0, output=1000)

This function calculates the number of steps you need to set and the number of frames that you will get in the MD simulation.

Arguments

t_simulation: The total simulation time in the specified unit.

unit: The unit of the simulation time. It can be "ns" (nan

timestep: The time step of the simulation in the specified unit.

output: The frequency of frame storage during the MD simulation.

LumberjackToolkit.write_frame! — Method

write_frame!(trajectory_pdb_file, atoms, frame

Write a frame of a trajectory in the temporary PDB trajectory file.

LumberjackToolkit.writepdb_trajectory — Method

writepdb_trajectory(pdbname::String, trajectory::String; selection="all", first=1, last=nothing, step=1)

Write a trajectory in a temporary PDB file. This is very useful when you want to visualize the edited trajectory in VMD. The MDLovoFit program uses this function to write the trajectory in a PDB file before the analysis.